Downgrade too-new C++ API to use C API, including…
…the usage of Gdk::GLTextureBuilder in OptGLArea to GdkGLTextureBuilder.

+ drive-by fix of an unused variable in pack.cpp
+ fix undefined references from only partially migrated func signatures.

+++ src/ui/popup-menu.cpp
@@ -97,7 +97,7 @@ static void popup_at(Gtk::Popover &popover, Gtk::Widget &widget,
 }
 
 void popup_at(Gtk::Popover &popover, Gtk::Widget &widget,
-              int const x_offset, int const y_offset)
+              double const x_offset, double const y_offset)
 {
     popup_at(popover, widget, x_offset, y_offset, 0, 0);
 }
